minor refactor to DRY up BOM
